Making Must gather function a generic one
